comments are added to all the public methods in the core

+++ Hotel.Core/Managers/UserManager.cs
@@ -13,8 +13,16 @@ namespace Hotel.Core
         {
             _userRepository = userRepository;
         }
-
+        /// <summary>
+        /// gets all <see cref="IUser"/> from the repository
+        /// </summary>
         public IList<IUser> UsersList => _userRepository.GetAll();
+        /// <summary>
+        /// this method gets the user async from email and password
+        /// </summary>
+        /// <param name="email"></param>
+        /// <param name="password"></param>
+        /// <returns> <see cref="IUser"/> if found and null if not</returns>
         public async Task<IUser?> GetUserFromEmailPassAsync(string email, string password)
         {
             if (email == null || password == null)
@@ -24,7 +32,16 @@ namespace Hotel.Core
             var user = await _userRepository.GetUserWithAuthAsync(email, password);
             return user;
         }
-
+        /// <summary>
+        /// This method creates a new user
+        /// </summary>
+        /// <remarks> if email is already used by another user <see cref="EmailAlreadyUsedException"/> is thrown</remarks>
+        /// <param name="firstName"></param>
+        /// <param name="lastName"></param>
+        /// <param name="email"></param>
+        /// <param name="password"></param>
+        /// <param name="isUserAdmin"></param>
+        /// <exception cref="EmailAlreadyUsedException"></exception>
         public void CreateNewUser(string firstName, string lastName, string email, string password, bool isUserAdmin)
         {
             var userId = Guid.NewGuid();
@@ -36,7 +53,18 @@ namespace Hotel.Core
             var user = new User(userId, NamingHelper.FixNameFormat(firstName), NamingHelper.FixNameFormat(lastName), email, isUserAdmin);
             _userRepository.CreateNewModel(user, password);
         }
-
+        /// <summary>
+        /// This method update a user with it id
+        /// <para> password can be given as null if it's not wished to be changed</para>
+        /// </summary>
+        /// <remarks>if user is not found <see cref="ArgumentNullException"/> is thrown</remarks>
+        /// <param name="userId"></param>
+        /// <param name="firstName"></param>
+        /// <param name="lastName"></param>
+        /// <param name="email"></param>
+        /// <param name="isUserAdmin"></param>
+        /// <param name="password"></param>
+        /// <exception cref="ArgumentNullException"></exception>
         public void UpdateUser(Guid userId, string firstName, string lastName, string email, bool isUserAdmin, string? password = null)
         {
             var user = UsersList.FirstOrDefault(x => x.Id == userId);
@@ -50,10 +78,24 @@ namespace Hotel.Core
             user.IsUserAdmin = isUserAdmin;
             _userRepository.UpdateModel(user,password);
         }
-
-        public void DeleteUser(IUser user) =>
+        /// <summary>
+        /// This methdo deletes a user
+        /// </summary>
+        /// <remarks>if user is not found <see cref="ArgumentNullException"/> is thrown</remarks>
+        /// <param name="user"></param>
+        public void DeleteUser(IUser user)
+        {
+            if (user == null)
+            {
+                throw new ArgumentNullException(nameof(user));
+            }
             _userRepository.DeleteModel(user);
-
+        }
+        /// <summary>
+        /// This method return a user from it's Id
+        /// </summary>
+        /// <param name="UserId"></param>
+        /// <returns><see cref="IUser"/> if found and null if not</returns>
         public IUser? GetUserFromId(Guid UserId) 
         {
             return UsersList.FirstOrDefault(x => x.Id == UserId);
